Separate use chat streaming
